# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "description": "Документация API вашего приложения",
        "contact": {
            "email": "support@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Основной API сервер"
        }
    ],
    "paths": {
...
        "/api/buildings": {
            "get": {
                "tags": [
                    "Здания"
                ],
                "summary": "Получить список всех зданий",
                "operationId": "704d1f7897ed7a3b8d29dacea66823ad",
                "responses": {
                    "200": {
                        "description": "Список зданий успешно получен",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Building"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Здания"
                ],
                "summary": "Создать новое здание",
                "operationId": "c8f8399919e01b1e4e2e2f6a9f1f5107",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "address"
                                ],
                                "properties": {
                                    "address": {
                                        "type": "string",
                                        "maxLength": 255,
                                        "example": "г. Москва, ул. Ленина, 10"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Здание успешно создано",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Building"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Ошибка валидации"
                    }
                }
            }
        },
        "/api/buildings/{id}": {
            "get": {
                "tags": [
                    "Здания"
                ],
                "summary": "Получить информацию о конкретном здании",
                "operationId": "7c2e9a5ec44b127b681717d1d5b91152",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID здания",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Информация о здании",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Building"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Здание не найдено"
                    }
                }
            },
            "put": {
                "tags": [
                    "Здания"
                ],
                "summary": "Обновить данные здания",
                "operationId": "0bd50384a833589efa669b51b8e1ff88",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID здания",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "address"
                                ],
                                "properties": {
                                    "address": {
                                        "type": "string",
                                        "example": "г. Санкт-Петербург, Невский пр., 12"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Здание успешно обновлено",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Building"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Здание не найдено"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Здания"
                ],
                "summary": "Удалить здание",
                "operationId": "db5360c892964cf865373a95f5da4633",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID здания",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Здание удалено"
                    },
                    "404": {
                        "description": "Здание не найдено"
                    }
                }
            }
        },
        "/api/buildings/{id}/organizations": {
            "get": {
                "tags": [
                    "Здания"
                ],
                "summary": "Получить список организаций, находящихся в здании",
                "operationId": "30f8994727996ee13a6408ec39a88ac9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID здания",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Организации успешно получены",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "building": {
                                            "type": "string",
                                            "example": "ул. Ленина, 10"
                                        },
                                        "organizations": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Organization"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/buildings/nearby": {
            "get": {
                "tags": [
                    "Здания"
                ],
                "summary": "Найти ближайшие здания по координатам",
                "description": "Возвращает список зданий в пределах указанного радиуса (по умолчанию 5 км)",
                "operationId": "7cf303a1ca50c70f4562626538463214",
                "parameters": [
                    {
                        "name": "lat",
                        "in": "query",
                        "description": "Широта",
                        "required": true,
                        "schema": {
                            "type": "number",
                            "example": 55.751244
                        }
                    },
                    {
                        "name": "lng",
                        "in": "query",
                        "description": "Долгота",
                        "required": true,
                        "schema": {
                            "type": "number",
                            "example": 37.618423
                        }
                    },
                    {
                        "name": "radius",
                        "in": "query",
                        "description": "Радиус поиска (в км)",
                        "required": false,
                        "schema": {
                            "type": "number",
                            "example": 5
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ближайшие здания успешно получены",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Building"
                                            }
                                        },
                                        "meta": {
                                            "properties": {
                                                "center": {
                                                    "properties": {
                                                        "latitude": {
                                                            "type": "number",
                                                            "example": 55.751244
                                                        },
                                                        "longitude": {
                                                            "type": "number",
                                                            "example": 37.618423
                                                        }
                                                    },
                                                    "type": "object"
                                                },
                                                "radius_km": {
                                                    "type": "number",
                                                    "example": 5
                                                },
                                                "found": {
                                                    "type": "integer",
                                                    "example": 3
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Activity": {
                "title": "Activity",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 5
                    },
                    "name": {
                        "type": "string",
                        "example": "Еда"
                    },
                    "parent_id": {
                        "type": "integer",
                        "example": null,
                        "nullable": true
                    },
                    "_lft": {
                        "type": "integer",
                        "example": 1
                    },
                    "_rgt": {
                        "type": "integer",
                        "example": 8
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "Building": {
                "title": "Здание",
                "description": "Здание, где расположена организация",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 10
                    },
                    "address": {
                        "type": "string",
                        "example": "г. Москва, ул. Ленина, 10"
                    },
                    "latitude": {
                        "type": "number",
                        "format": "float",
                        "example": 55.7558
                    },
                    "longitude": {
                        "type": "number",
                        "format": "float",
                        "example": 37.6173
                    }
                },
                "type": "object"
            },
            "Organization": {
                "title": "Организация",
                "description": "Модель организации",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "ООО Ромашка"
                    },
                    "building_id": {
                        "type": "integer",
                        "example": 3
                    },
                    "phones": {
                        "type": "array",
                        "items": {
                            "type": "string",
                            "example": "+7 (900) 123-45-67"
                        }
                    },
                    "activities": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Activity"
                        }
                    },
                    "building": {
                        "$ref": "#/components/schemas/Building"
                    }
                },
                "type": "object"
            }
        }
    },
...
}